Multiple Variogram Fitting

This function returns experimental and fitted variogram information for multiple variables. For each variable autofitVariogram is used. The function supports spatial data in simple features (sf), SpatialPointsDataFrame(sp) or data.frame. If the input data is in `data.frame`, the X, and Y should be present in the data as variables.

multiple_variogram(data, coords = c("X", "Y"), length = 99)

Arguments

data

input data set one of sp, sf or data.frame with X and Y as variables

coords

(vector) pair of coordinates if data type is aspatial

length

(integer) the length of sequence used to produced fitted data.

Value

A data frame with combined information from all variables as response is returned. The `data.frame` contains the experimental variogram information, the fitted variogram model information. The sums of squares (sserr) between the sample variogram dropped, For further information please refer to autofitVariogram or fit_variogram

References

automatp: A index page for automap package on [CRAN](https://cran.r-project.org)

Hiemstra, P.H., Pebesma, E.J., Twenhofel, C.J.W. and G.B.M. Heuvelink, 2008. Real-time automatic interpolation of ambient gamma dose rates from the Dutch Radioactivity Monitoring Network. Computers & Geosciences.[DOI:](http://dx.doi.org/10.1016/j.cageo.2008.10.011)

See also

Examples

## Read data
lcdat<- landcover

# subset data

lcdat <- lcdat[, c("MBLU", "MNDVI", "MNIR", "MPC1", "MPC2", "MPC23", "MRED", "MSAVI")]

# fit variograms

fmv<- multiple_variogram(data = lcdat,coords = NULL,length = 99 )
#> [1] "running semivariance analysis for variable : MBLU"
#> [1] "running semivariance analysis for variable : MNDVI"
#> [1] "running semivariance analysis for variable : MNIR"
#> [1] "running semivariance analysis for variable : MPC1"
#> [1] "running semivariance analysis for variable : MPC2"
#> [1] "running semivariance analysis for variable : MPC23"
#> [1] "running semivariance analysis for variable : MRED"
#> [1] "running semivariance analysis for variable : MSAVI"

if (FALSE) {
data("landcover")
fit_multiple_variogram(data = landcover,coords = NULL, length = 99)
}