Spatial Thinning of Multi-class Point Data

The function is a wrapper around spatial thinning thin function from spThin package. It makes spatial thinning easy for multi-class input data set. This function is handy and flexible: can support xy coordinates, spatialPoints, or simple features. It assumes that data are in projected planar coordinates system; preferably Universal Transverse Mercator (UTM). There is no restrictions on number of columns in the data set as long as data is in the required format.

spatial_thinning_class(
  data,
  spatial = FALSE,
  coords = c("X", "Y"),
  distance,
  reps = 1,
  class = "class"
)

Arguments

data

input data set one of sp, sf or data.frame with X and Y as variables

spatial

(logical) if variable is spatial or aspatial

coords

(vector) pair of coordinates if data type is aspatial

distance

minimum distance between two points of same class

reps

(integer) replication

class

multi-class variable or strata variable. For each class thinning will be applied based on specified distance

Value

data.frame

Details

Spatial thinning expect the data in the projected coordinate system, the distance value is the threshold which limits the number of observations. All the observations, in a particular class, smaller than the distance thresholds are deleted. If there are more than two classes, minimum distance between observations from each class.

References

Aiello-Lammens, M. E., Boria, R. A., Radosavljevic, A. , Vilela, B. and Anderson, R. P. (2015). spThin: an Rpackage for spatial thinning of species occurrence records for use in ecological niche models. Ecography, 38:541-545. URL https://onlinelibrary.wiley.com/doi/10.1111/ecog.01132.

See also

Examples

## load data from the package
tdat<- landcover

# Run spatial thinning (sf format)
stc<- spatial_thinning_class(data = tdat,spatial = TRUE,
coords = NULL,distance = 10000,reps = 1,class = "Class_name")
#> [1] "running spatial random thinning of class: 1"
#> [1] "running spatial random thinning of class: 2"
#> [1] "running spatial random thinning of class: 3"
#> [1] "running spatial random thinning of class: 4"
#> [1] "running spatial random thinning of class: 5"
#> [1] "running spatial random thinning of class: 6"
#> [1] "running spatial random thinning of class: 7"

### compare the number of samples from original data to the thinned data
## Original data
library(dplyr)
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
tdat %>% group_by(Class_name) %>% summarize(n = n())
#> Simple feature collection with 7 features and 2 fields
#> Geometry type: MULTIPOINT
#> Dimension:     XY
#> Bounding box:  xmin: 390919.6 ymin: 3119625 xmax: 460847.2 ymax: 3168895
#> Projected CRS: NAD83 / UTM zone 14N
#> # A tibble: 7 x 3
#>   Class_name     n                                                      geometry
#>        <dbl> <int>                                              <MULTIPOINT [m]>
#> 1          1   115 ((398855.9 3120719), (398890.5 3120817), (398928.3 3120852),~
#> 2          2   154 ((397279 3119856), (398273.7 3122494), (398279 3122515), (40~
#> 3          3   770 ((391209.4 3162493), (391254.4 3162353), (391283.9 3162566),~
#> 4          4   314 ((390919.6 3131268), (390940.3 3120156), (390944.1 3131090),~
#> 5          5   451 ((391205.2 3162390), (391608.8 3129863), (391648.1 3129799),~
#> 6          6    91 ((391884.2 3145948), (391949.5 3143032), (392085.6 3120606),~
#> 7          7    27 ((394204.2 3133275), (397727 3166709), (399156.2 3133834), (~

## Thinned data
stc %>% group_by(Class_name) %>% summarize(n = n())
#> # A tibble: 7 x 2
#>   Class_name     n
#>        <dbl> <int>
#> 1          1     9
#> 2          2    13
#> 3          3    28
#> 4          4    28
#> 5          5    24
#> 6          6    19
#> 7          7     7

##  spatial thinning using meuse data
# load library and data
library(sp)
data(meuse)

# apply thinning

mthin<- spatial_thinning_class(data = meuse, spatial = FALSE,
         coords = c("x", "y"),
         distance = 1000,
         reps = 2, class = "ffreq"
         )
#> [1] "running spatial random thinning of class: 1"
#> [1] "running spatial random thinning of class: 2"
#> [1] "running spatial random thinning of class: 3"

if (FALSE) {
data<- landcover
spatial_thinning_class(data = landcover, spatial = TRUE, coords = NULL,
distance = 10000, reps = 2, class = "Class_name")
}