Land Use Land Cover (LULC) Data for Classification Problem

The land use land cover data were generated using heads-up (onscreen) digitization using National Agriculture Imagery Program (NAIP) digital orthoimagery data acquired in 2018.

Usage

landcover

Format

A data frame with 1922 rows and 32 variables:

MBLU

double mean Blue

MDIS

double Mean Dissimilarity

MENT

double mean Entropy

MGRN

double Mean Green

MHOM

double Mean Homogeneity

MNDSI

double Mean Normalized Difference Soil Index

MNDVI

double Mean Normalized Difference Vegetation Index

MNDWI

double Mean Normalized Difference Water Index

MNIR

double Mean Near infrared

MPC1

double Mean Principal Component Axis 1

MPC2

double double Mean Principal Component Axis 2

MPC23

double double Mean Principal Component Axis 3

MRED

double Mean Red

MSAVI

double Mean Soil Adjusted Vegetation Index

MSTD

double Mean Standard Deviation

SAVG

double Standard deviation of AVG

SAVI

double Standard deviation of SAVI

SBLU

double Standard deviation of Blue

SDIS

double Standard deviation of Dissimilarity

SENT

double Standard deviation of Entropy

SGRN

double Standard deviation of Green

SHOM

double Standard deviation of Homogeneity

SHPI

double Shape Index

SNDSI

double Standard deviation of NDSI

SNDVI

double Standard deviation of NDVI

SNIR

double Standard deviation of Near infrared

SPC1

double Standard deviation of PCA1

SPC2

double Standard deviation of PCA2

SPC3

double Standard deviation of PCA3

SRED

double Standard deviation of Red

SSTD

double Standard deviation of Standard Deviation

geometry

list Geometry Information

Class_name

double land use land cover classes 1 through 7

Details

Slightly doctored training data set of land use land cover. Original data points were generated using heads-up (on-screen) digitization. The data has 31 predictor variables and 1 dependent multi class (Class_name) variable with 1922 records.